[stats] make lock-free mutations
Summary: In multicore mode, analysis workers will contend on the single mutex protecting the whole stats record in `Stats`. We avoid this by reworking the ppx code generator and the stats code so that they operate on a record without `mutable` fields, but fields of type `'a Atomic.t`. These fields are updated using a compare-and-set loop, which should reduce contention.

Reviewed By: dulmarod

Differential Revision:
D82625368

Privacy Context Container: L1208441

fbshipit-source-id: 69b898faaa76d9074f95b83230d7d6dccd7bef5a