[semdiff] implement diff algorithm in ocaml
Summary:
this diff starts to migrate python code used by semdiff to OCaml. This will make it easier to test the implementation and prepare the tool for potential new features later on.

The first part to be migrated is the diff algorithm. We implement the `get_diff` function in OCaml and we define a tree in OCaml using this minimal type:
```
type ast_node =
  | Dict of (string * ast_node) list
  | List of ast_node list
  | Str of string
  | Num of int
  | Bool of bool
  | Null
```

Reviewed By: ngorogiannis

Differential Revision:
D83487341

Privacy Context Container: L1208441

fbshipit-source-id: 6bc527125708b87c9f7c24b94a4e9fdc0833b608