[semdiff] reduce noise in output
Summary: we don't need to show repeated lines when they just change position in the diff. Instead we omit them and show only content which was actually added/removed. The next step is to add line numbers to output, which are hidden for the moment.

Reviewed By: ngorogiannis

Differential Revision:
D84354957

Privacy Context Container: L1208441

fbshipit-source-id: 7c66fa39abd03eafc64da8c9a07d6fafdc707939