[swift] Adding example of not retain cycle with weak attribute
Summary:
Adding a weak attribute to the translation that shows the changes in the sil file and also that the retain cycle disappears.  This is because we need to model or translate away the function `swift_weakAssign` in the setter.

We will also need to handle `swift_weakLoadStrong` in the getter.

Reviewed By: ngorogiannis

Differential Revision: D84623768

fbshipit-source-id: 51de2a431355321a61db5934468a9783d9831cd7