[infinite-loop] record path stamp at each loop header and each iteration
Summary: this the final diff. Now we have an other way to track infinite loops. The logic is the same but we don't need to use special widening now. It should help with understanding the analysis better (specially) for HTML, and also make the analysis a bit faster.

Reviewed By: ngorogiannis

Differential Revision:
D84464725

Privacy Context Container: L1208441

fbshipit-source-id: 83606310b4b9f13c81170bc09862815f3e402f93