[swift] Adding the weak field to field names in the tenv
Summary: Adding the weak field to field names in the tenv. Still need to add it to other field names in the program. For context, in the context of retain cycles for Objective-C we added this field is_weak in the actual fieldname because sometimes during the analysis we do not have the correct tenv to look up this value from. It's also more efficient to look this up directly in the name.

Reviewed By: ngorogiannis

Differential Revision: D84716995

fbshipit-source-id: 3c260722f19026c00970cc058b058f27ef26c9c1