[swift] Translate away weak specific builtins
Summary:
`swift_weakInit` is just initialising the memory of weak pointers, not needed for us, so we skip it.  "`swift_weakInit` initializes a weak reference slot in memory for Swift objects.
It ensures the runtime can track and zero out weak references when the referenced object is deallocated."

`n4 = swift_weakLoadStrong(n3) `, we translate it as `n4 = n3` , we can also skip it. "Loads a weak reference and returns a strong reference if alive".

Reviewed By: davidpichardie

Differential Revision: D84819245

fbshipit-source-id: aa91ffcdb22e25f5f336aa5d5faa8401e89bec0d