[swift] Translate swift_weakAssign as a store
Summary:
We translate swift_weakAssign as a store instruction.

Replace swift_weakAssign with a store instruction. We do not add dereference to the first argument because we are flattenning the structure of weak pointers to be just like normal pointers in infer, whilst in llvm the structures is `field_2: *swift::weak}, type swift::weak = {field_0: *ptr_elt} `.

Reviewed By: ngorogiannis

Differential Revision: D84833013

fbshipit-source-id: 74ab49b7271dda4b700c88f36c968221389aedc7