[swift] Translate the type of optionals better
Summary:
The type of optionals
`%TSiSg = type <{ [8 x i8], [1 x i8] }>` was translated as `int[]` but this is a struct representing Optional<Int> in Swift.
`[8 x i8]`: 8 bytes for the Int value (payload)
`[1 x i8]`: 1 byte for the "some/none" tag (discriminator)

We now translate those types as Int and Bool respectively. In Textual we do not have a Bool type though, so it's Int.

Reviewed By: davidpichardie

Differential Revision: D85051637

fbshipit-source-id: f4132e7c5b1e841b62712f86c6d994cc64884a55