[swift] Improve translation of phi nodes
Summary:
In llair the phi nodes are translated by the move instructions, which is also overloaded to translate other instructions. Whilst not having a load is fine in other cases, it's not fine here. We can see in our example that the phi node gives us a variable `n5` that is either `0` or `var0`, and of type int. For this typing to work, we need to dereference the variable, so that `n5` is the actual int.

So, we add a new instruction to llair, `MovePhi`, that we use only for this case, and in the translation to Textual we add a dereference. This solves the remaining typing problem.

Reviewed By: davidpichardie

Differential Revision: D85241717

fbshipit-source-id: bd7fc870890e7c45129260ee048a3a59e05123bc