[textual] optimize replace_dot_with_2colons
Summary:
This function is used a lot during capture, and seems to consume 5-10% on some Swift captures.
It turns out that the function used for string replacement creates an auxiliary data structure (for the KMP algorithm) and then throws it away every time. We optimise this by saving that data structure and reusing it.

Reviewed By: davidpichardie

Differential Revision:
D85441171

Privacy Context Container: L1208441

fbshipit-source-id: 744e004d577f49b884b3523d070bf6073c04136a