[swift] Update the struct_map with the plain names using the new type name structure
Summary:
This is a breaking change that I still need to split in various diffs because it's too big. We need to match mangled and plain type names everywhere. Here we do it in the struct_map.
Still to do: fields and the rest of the usages in the code.

Reviewed By: ngorogiannis

Differential Revision: D85878084

fbshipit-source-id: 9206d77dcf05963d9e647d83961ac420b3558a1e