[llvm] Updating method signatures after having obtained the classes the methods belong to and their offsets
Summary: We need the function signatures first to obtain the types, then we can compute the method's offsets, but then we need to update the method signatures again with the correct name and offset. Still to do is to update the types in the method signatures that currently still only include the plain class names.

Reviewed By: ngorogiannis

Differential Revision:
D85961964

Privacy Context Container: L1208441

fbshipit-source-id: fd9ad76ecbe34f5f628dc64b983c05ddf1814481