[swift] Use the return type of methods for type inference
Summary:
We add a `proc_map` to the proc_state to look up the signature of methods that we are calling. We also add a new map for id types, we already had one but this is specific for the ids that come from a move and for which we need to add a load instruction. We add the type only when it's not `ptr_elt` that won't help in finding the correct method.

In a next step, we will use those types gained to lookup the correct method for a call directly in the frontend when need it.

Reviewed By: ngorogiannis

Differential Revision: D86772343

fbshipit-source-id: 868dd0a302dbb148c7d4c51539a88471088934e7