[swift] Use local with debug name as formal
Summary:
This helps with getting debug names for the parameters of the functions. In llvm the debug names are there instead for the local mirrors of the parameters. We now remove the step of copying the parameters locally, and use the debug names for the parameters instead.

This also helps with the fact that we were doing a load on potentially nil parameters (optionals) in those instructions for copying the parameters.

Reviewed By: ngorogiannis

Differential Revision: D87086137

fbshipit-source-id: 694509270deb6bd39d78b7ba7bd4bb4f031d4568