[swift] Fix typing error
Summary: Previously we removed a local copy of the formals, and use the local variable with the debug name as a formal directly. In those cases we were getting type error sometimes when the signature type was int, because internally int is a pointer to a struct. Now, we use the local type for the formal in case of such a contradiction. In general, we prefer signature types because they are more clear.

Reviewed By: ngorogiannis

Differential Revision: D87784294

fbshipit-source-id: 4b6eb8f06ded18181b319657c51400a2e5144f92