[textualsil] push error monad inside Module.to_sil
Summary:
Since none of the immediate callees of `Module.to_sil` return a `result`, but instead raise exceptions on error, a single error means we completely fail the translation of the module.

This diff wraps each callee so that an exception is converted to an error result. We then collect all errors sequentially, without terminating the translation upon error. This means we (conditionally, on a CLI flag) do a best effort translation, while logging errors.

Reviewed By: davidpichardie

Differential Revision:
D87916192

Privacy Context Container: L1208441

fbshipit-source-id: 9447aa86cd2f7ad41fe5d578ca4a62301dd97705