[swift] Do not store empty signature types
Summary: We noticed in an example that when a signature type was the empty string, it would break the signature types. The cause of this is that is_substring always applies to the empty string. So, now we avoid adding those empty strings to the hash set of signature types.

Reviewed By: ngorogiannis

Differential Revision:
D87868436

Privacy Context Container: L1208441

fbshipit-source-id: d3380888a9bd089dde8ab1b507499f2e4afb00eb