[rust] Textual Translation: Assignment and constants (#1957)
Summary:
This PR introduces initial translation of rust programs.
This implements the data structures needed to run a minimal function with constants and assign statements.

The main changes / additions in this pr are as follows:
- `infer/src/integration/Rust.ml` : Support for translating and verifying the textual files..
- `infer/src/rust/RustMir2Textual.ml`: The main addition where the actual translation from rust to textual happends.
- `infer/src/rust/unit/`: The test folder, the folder is structed as followings:
    - `RustMir2TextualTest*.ml`: Contain the expected output of the tests

The tests are based on the intermediate representation (MIR) of the rust programs. To see the intermediate representation you can run copy the code to a rust file and run it  with `charon rustc --print-ullbc -- <file.rs>`
The output would then look something like this:
```
# Final ULLBC before control-flow reconstruction:

// Full name: example::main
fn main()
{
    let 0: (); // return
    let x@1: i32; // local
    let 2: (); // anonymous local

    bb0: {
        storage_live(x@1);
        x@1 := const (42 : i32);
        nop;
        0 := ();
        storage_dead(x@1);
        0 := ();
        return;
    }
}
```

Pull Request resolved: https://github.com/facebook/infer/pull/1957

Reviewed By: ngorogiannis

Differential Revision: D86079192

Pulled By: davidpichardie

fbshipit-source-id: 7b400f6be4b52f28853bef842ea945d02cbe7635