[swift] Resolve more functions at compile time
Summary:
The goal is to keep track of more types that we are sure of to resolve calls in the frontend. That saves time in Pulse execution and makes the IR more readable.

The calls we create here are virtual, so we can still resolve at run time if there is inheritance. That's not implemented yet.

Reviewed By: davidpichardie

Differential Revision: D87979166

fbshipit-source-id: 4414665e716ba01e7a277e3fc216d36952557c1c