[merge] avoid race condition on stats file
Summary:
When merging stats files, the main process may race with the type environment merger process on the stats file, as the latter may issue log messages itself.

We make the merge safer by:
1. open/close the main stats file only once instead of once per merged stats file
2. wait until the tenv merger process has terminated before merging all the files.
3. catch Json parse exceptions and skip

This will make the merge a bit less efficient. If this turns out to be a problem we can break this process into reading the stats files, which can be done in parallel with tenv merging, and writing out which will be done after.

Reviewed By: martintrojer

Differential Revision: D88070126

fbshipit-source-id: fb70e7ab8423d39e357f247396ef8e20f3f56068