[swift] pp full name in verbose mode, not just plain
Summary:
It is possible for procedures to have the same plain name but a different mangled one. Using only the plain name to generate DB keys will therefore lead to clashes.

Here we use the facility of `Mangled.to_string_full` to compose a key of the form `plain{mangled}` when a mangled name is available.

Reviewed By: dulmarod

Differential Revision: D88075476

fbshipit-source-id: b5a2132384b724b7d36a7367eed4a0e0a0c03289