[swift] Translate closures using the Closure expression
Summary: Using the closure expression to allow for specialisation like we do in Objective-C. The previous version with the function name as strings doesn't allow that. The model needs to be fixed in the follow up diff. We cannot do the Textual transformation like in other languages because the transformation is already done in llvm.

Reviewed By: davidpichardie

Differential Revision:
D88656252

Privacy Context Container: L1208441

fbshipit-source-id: 8f356e9c4fdce9d5ebbfb9e8fdb3e7fc9e351bc3