[swift] Test with captures variables and without.
Summary:
Adding another example of a closure call that needs specialisation that shows that specialisation works correctly when there are captured variables: in testWithClosure we return 20. However, when there are no captured variables, specialisation now fails with mismatched arguments (example2 when called in testWithClosure2). This is because the last argument of the call is the structure of captured variables, which is not needed in closures that don't capture variables.
We plan to address this by calling the closure again without the last argument after a failure.

Reviewed By: ngorogiannis

Differential Revision: D88863879

fbshipit-source-id: 8a1b7f605030f2c0c458e5c8fd3f6e9a4ebf6a79