[ci] add basic workflow testing the dev setup (#1981)
Summary:
When working with the open-source code of Infer, the `build-infer.sh` is sensitive to the version of OCaml being used as well as the platform and the different options passed to the script, e.g. `--no-opam-lock`, `--only-setup-opam`, `--user-opam-switch` or the default without option. Several combinations of these are not working.

The goal of this new workflow is to:
- document the combination that are known to be working
- prevent regressions for the development builds
- extend the matrix of possibilities when new combinations are becoming functional
- test several other commands used for development like `make devsetup`, `make -j -C infer/src check`, etc.

Pull Request resolved: https://github.com/facebook/infer/pull/1981

Reviewed By: ngorogiannis

Differential Revision: D88951009

Pulled By: dulmarod

fbshipit-source-id: b016ba06f09c57b540c3cba90a30964373873a09