[swift] Fix specialisation of closures when there are no captured variables
Summary: If the call is Unsat, it could be because of mismatched arguments, because the specialised closure doesn't capture any variables, and so the captured argument is not needed. In this case, we try the call again, by removing the last argument, which will be null.

Reviewed By: davidpichardie

Differential Revision:
D88950235

Privacy Context Container: L1208441

fbshipit-source-id: 31cd8212f4d6dce05b6640bb5c2ddff12fa8adcf