[llvm] memoize mangled_map
Summary:
When searching for a struct type name based on the mangled name, we do a linear scan of all types (in `struct_name_of_mangled_name`). This can be costly, because this operation is called in several places, including `to_textual_typ`.

We memoize the map from mangled names to struct names *after* the struct names have stabilised, that is, after `translate_llair_function_signatures` in `init_module_state`. We then use the memoized map to look up struct names efficiently.

We also change the exposed interface of a few functions so as to avoid providing the mangled map via an error-prone optional argument, so we have `to_annotated_textual_typ_without_mangled_map` and `to_textual_typ_without_mangled_map` which are used exactly once each.

Reviewed By: davidpichardie

Differential Revision: D88933130

fbshipit-source-id: 88306795166d8361feaea16c116e0df9fcf0ef5d