[swift] Find more field names
Summary:
This diff improves the algorithm for finding field names in getters, by taking into account another instruction that appears in getters: move, with the field selector as an expression to be moved to a variable.
Moreover, when we don't find the type as the type of the expression, we can find it as the plain name type of the argument of the getter, since we know the variable is the argument.

Reviewed By: davidpichardie

Differential Revision: D89299911

fbshipit-source-id: 4ee02226505aebb32a89e82a1629bd2afea726b5