[swift] Fix translation of move_phi
Summary: When we translate phi nodes, in llair this is a move_phi, and, whilst in normal move, we do not deref variables, here it's necessary. It was needed in an example for optional types. Now we see another example, when adding deref is a problem, because the translation adds it again later leading to a typechecking error. Now we keep track of that the deref has been already added in the move, and we don't add it later. With this approach both examples work correctly.

Reviewed By: davidpichardie

Differential Revision:
D89551549

Privacy Context Container: L1208441

fbshipit-source-id: 7e0c80ef28f6d58c075467b28d89bbf99455fa5f